Replace test of drs localization PAPI action
